rename storage to object_stores
