+++ crates/iceberg-catalog/src/api/management/v1/warehouse.rs
@@ -1,7 +1,7 @@
 use crate::api::management::v1::{ApiServer, DeletedTabularResponse, ListDeletedTabularsResponse};
 use crate::api::{ApiContext, Result};
 use crate::request_metadata::RequestMetadata;
-pub use crate::service::storage::{
+pub use crate::service::object_stores::{
     AzCredential, AzdlsProfile, S3Credential, S3Profile, StorageCredential, StorageProfile,
 };
