+++ crates/iceberg-catalog/src/service/mod.rs
@@ -4,8 +4,8 @@ pub mod config;
 pub mod contract_verification;
 pub mod event_publisher;
 pub mod health;
+pub mod object_stores;
 pub mod secrets;
-pub mod storage;
 pub mod tabular_idents;
 pub mod task_queue;
 pub mod token_verification;
