+++ crates/iceberg-catalog/src/service/object_stores/mod.rs
